pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "megalodon-enwik8-jax"
version = "0.1.0"
description = "JAX port of MEGALODON character-level language modeling on enwik8"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "Peter Szemraj", email = "peterszemraj+megalodon@gmail.com" },
]
keywords = ["jax", "megalodon", "language-model", "enwik8", "character-level"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "jax>=0.4.30",
    "equinox>=0.12.0",
    "optax>=0.2.0",
    "megalodon-jax==0.1.0",
    "pyyaml>=6.0",
    "numpy>=1.26.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
checkpoint = [
    "orbax-checkpoint>=0.5.0",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.4.0",
]

[project.scripts]
train-megalodon = "megalodon_enwik8_jax.cli:train_main"
infer-megalodon = "megalodon_enwik8_jax.cli:inference_main"

[tool.hatch.build.targets.wheel]
packages = ["src/megalodon_enwik8_jax"]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = ["vendor"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]
ignore = ["E501", "F722", "SIM108"]  # F722: jaxtyping; SIM108: ternary style

[tool.ruff.lint.isort]
known-first-party = ["megalodon_enwik8_jax"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short"
